move collection struct into collection.rs
