+++ lib/collection/src/operations/conversions.rs
@@ -1,19 +1,17 @@
-use crate::config::{CollectionParams, WalConfig};
-use crate::operations::config_diff::{HnswConfigDiff, WalConfigDiff};
+use crate::config::{CollectionConfig, CollectionParams, WalConfig};
+use crate::operations::config_diff::{HnswConfigDiff, OptimizersConfigDiff, WalConfigDiff};
 use crate::operations::point_ops::PointsSelector::PointIdsSelector;
 use crate::operations::point_ops::{
     Batch, FilterSelector, PointIdsList, PointStruct, PointsSelector,
 };
-use crate::operations::types::{CollectionStatus, OptimizersStatus, UpdateStatus};
-use crate::{
-    CollectionConfig, CollectionInfo, CountResult, OptimizersConfig, OptimizersConfigDiff, Record,
-    UpdateResult,
-};
+use crate::operations::types::{CollectionInfo, CollectionStatus, CountResult, OptimizersStatus, Record, UpdateResult, UpdateStatus};
+
 use api::grpc::conversions::{payload_to_proto, proto_to_payloads};
 use itertools::Itertools;
 use std::collections::HashMap;
 use std::num::NonZeroU32;
 use tonic::Status;
+use crate::optimizers_builder::OptimizersConfig;
 
 impl From<api::grpc::qdrant::HnswConfigDiff> for HnswConfigDiff {
     fn from(value: api::grpc::qdrant::HnswConfigDiff) -> Self {
