+++ lib/collection/src/operations/payload_ops.rs
@@ -3,7 +3,8 @@ use segment::types::{Filter, Payload, PayloadKeyType, PointIdType};
 use serde;
 use serde::{Deserialize, Serialize};
 
-use crate::{hash_ring::HashRing, ShardId};
+use crate::hash_ring::HashRing;
+use crate::shard::ShardId;
 
 use super::{split_iter_by_shard, OperationToShard, SplitByShard};
