+++ lib/collection/src/operations/snapshot_ops.rs
@@ -1,10 +1,10 @@
-use crate::CollectionResult;
 use api::grpc::conversions::date_time_to_proto;
 use chrono::NaiveDateTime;
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 use std::path::Path;
 use std::time::SystemTime;
+use crate::operations::types::CollectionResult;
 
 #[derive(Debug, Deserialize, Serialize, JsonSchema, Clone)]
 pub struct SnapshotDescription {
