+++ lib/collection/src/shard/conversions.rs
@@ -2,7 +2,6 @@ use crate::operations::payload_ops::{DeletePayload, SetPayload};
 use crate::operations::point_ops::PointInsertOperations;
 use crate::operations::CreateIndex;
 use crate::shard::remote_shard::RemoteShard;
-use crate::CollectionResult;
 use api::grpc::conversions::payload_to_proto;
 use api::grpc::qdrant::points_selector::PointsSelectorOneOf;
 use api::grpc::qdrant::{
@@ -14,6 +13,7 @@ use api::grpc::qdrant::{
 };
 use segment::types::{Filter, PointIdType};
 use tonic::Status;
+use crate::operations::types::CollectionResult;
 
 pub fn internal_upsert_points(
     point_insert_operations: PointInsertOperations,
