+++ lib/collection/src/shard/local_shard_operations.rs
@@ -1,10 +1,6 @@
 use crate::collection_manager::segments_searcher::SegmentsSearcher;
-use crate::operations::types::{CollectionStatus, OptimizersStatus, UpdateStatus};
+use crate::operations::types::{CollectionInfo, CollectionResult, CollectionStatus, CountRequest, CountResult, OptimizersStatus, PointRequest, Record, SearchRequest, UpdateResult, UpdateStatus};
 use crate::update_handler::{OperationData, UpdateSignal};
-use crate::{
-    CollectionInfo, CollectionResult, CollectionUpdateOperations, CountRequest, CountResult,
-    LocalShard, PointRequest, Record, SearchRequest, ShardOperation, UpdateResult,
-};
 use async_trait::async_trait;
 use itertools::Itertools;
 use segment::types::{
@@ -15,6 +11,9 @@ use std::collections::HashMap;
 use std::sync::Arc;
 use tokio::runtime::Handle;
 use tokio::sync::oneshot;
+use crate::operations::CollectionUpdateOperations;
+use crate::shard::local_shard::LocalShard;
+use crate::shard::ShardOperation;
 
 #[async_trait]
 impl ShardOperation for &LocalShard {
