+++ lib/collection/src/shard/mod.rs
@@ -6,16 +6,20 @@ pub mod proxy_shard;
 pub mod remote_shard;
 pub mod shard_config;
 
+use std::collections::HashMap;
 use crate::shard::proxy_shard::ProxyShard;
 use crate::shard::remote_shard::RemoteShard;
-use crate::{
-    CollectionInfo, CollectionResult, CollectionUpdateOperations, CountRequest, CountResult,
-    LocalShard, PeerId, PointRequest, Record, SearchRequest, UpdateResult,
-};
+use std::path::{Path, PathBuf};
+
+use crate::operations::types::{CollectionError, CollectionInfo, CollectionResult, CountRequest, CountResult, PointRequest, Record, SearchRequest, UpdateResult};
+use crate::shard::local_shard::LocalShard;
 use async_trait::async_trait;
 use segment::types::{ExtendedPointId, Filter, ScoredPoint, WithPayload, WithPayloadInterface};
 use std::sync::Arc;
 use tokio::runtime::Handle;
+use tonic::transport::Uri;
+use api::grpc::transport_channel_pool::TransportChannelPool;
+use crate::operations::CollectionUpdateOperations;
 
 pub type ShardId = u32;
 
@@ -90,4 +94,59 @@ pub trait ShardOperation {
         with_payload: &WithPayload,
         with_vector: bool,
     ) -> CollectionResult<Vec<Record>>;
+}
+
+pub const HASH_RING_SHARD_SCALE: u32 = 100;
+
+pub type CollectionId = String;
+
+pub type PeerId = u64;
+
+#[derive(Clone)]
+pub struct ChannelService {
+    pub id_to_address: Arc<parking_lot::RwLock<HashMap<u64, Uri>>>,
+    pub channel_pool: Arc<TransportChannelPool>,
+}
+
+impl ChannelService {
+    pub fn new(
+        id_to_address: Arc<parking_lot::RwLock<HashMap<u64, Uri>>>,
+        channel_pool: Arc<TransportChannelPool>,
+    ) -> Self {
+        Self {
+            id_to_address,
+            channel_pool,
+        }
+    }
+}
+
+impl Default for ChannelService {
+    fn default() -> Self {
+        Self {
+            id_to_address: Arc::new(Default::default()),
+            channel_pool: Arc::new(Default::default()),
+        }
+    }
+}
+
+pub struct ShardTransfer {
+    pub from: PeerId,
+    pub to: PeerId,
+}
+
+pub fn shard_path(collection_path: &Path, shard_id: ShardId) -> PathBuf {
+    collection_path.join(format!("{shard_id}"))
+}
+
+pub async fn create_shard_dir(
+    collection_path: &Path,
+    shard_id: ShardId,
+) -> CollectionResult<PathBuf> {
+    let shard_path = shard_path(collection_path, shard_id);
+    tokio::fs::create_dir_all(&shard_path)
+        .await
+        .map_err(|err| CollectionError::ServiceError {
+            error: format!("Can't create shard {shard_id} directory. Error: {}", err),
+        })?;
+    Ok(shard_path)
 }
