+++ lib/collection/src/shard/shard_config.rs
@@ -1,8 +1,9 @@
-use crate::{CollectionResult, PeerId};
 use std::path::{Path, PathBuf};
 
 use segment::common::file_operations::{atomic_save_json, read_json};
 use serde::{Deserialize, Serialize};
+use crate::operations::types::CollectionResult;
+use crate::shard::PeerId;
 
 pub const SHARD_CONFIG_FILE: &str = "shard_config.json";
