+++ lib/collection/tests/common/mod.rs
@@ -2,10 +2,11 @@ use collection::config::{CollectionConfig, CollectionParams, WalConfig};
 use collection::operations::types::CollectionError;
 use collection::optimizers_builder::OptimizersConfig;
 use collection::shard::collection_shard_distribution::CollectionShardDistribution;
-use collection::{ChannelService, Collection, CollectionId};
 use segment::types::Distance;
 use std::num::NonZeroU32;
 use std::path::Path;
+use collection::collection::Collection;
+use collection::shard::{ChannelService, CollectionId};
 
 /// Test collections for this upper bound of shards.
 /// Testing with more shards is problematic due to `number of open files problem`
