+++ lib/collection/tests/snapshot_test.rs
@@ -1,10 +1,10 @@
 use crate::common::TEST_OPTIMIZERS_CONFIG;
 use collection::config::{CollectionConfig, CollectionParams, WalConfig};
 use collection::shard::collection_shard_distribution::CollectionShardDistribution;
-use collection::shard::Shard;
-use collection::{ChannelService, Collection};
+use collection::shard::{ChannelService, Shard};
 use segment::types::Distance;
 use std::num::NonZeroU32;
+use collection::collection::Collection;
 
 mod common;
