+++ lib/storage/src/content_manager/collection_meta_ops.rs
@@ -2,11 +2,11 @@ use crate::content_manager::shard_distribution::ShardDistributionProposal;
 use collection::{
     operations::config_diff::{HnswConfigDiff, OptimizersConfigDiff, WalConfigDiff},
     shard::ShardId,
-    CollectionId, PeerId,
 };
 use schemars::JsonSchema;
 use segment::types::Distance;
 use serde::{Deserialize, Serialize};
+use collection::shard::{CollectionId, PeerId};
 
 // *Operation wrapper structure is only required for better OpenAPI generation
